+++ Mods ChoGGi/Change Object Colour/metadata.lua
@@ -17,5 +17,5 @@ Use the "Default" checkbox to reset colours.
 
 You can also use Shift+F6 or Ctrl+F6 for random colours and default colours.
 
-Part of ECM.]],
+Part of Expanded Cheat Menu.]],
 })