+++ Mods ChoGGi/Fix Remove Blue Yellow Grid Marks/metadata.lua
@@ -17,5 +17,5 @@ If you keep getting transport routes stuck then use this:
 https://steamcommunity.com/sharedfiles/filedetails/?id=1526896729
 
 
-Part of ECM.]],
+Part of Expanded Cheat Menu.]],
 })