+++ Mods ChoGGi/Terraformer/metadata.lua
@@ -23,5 +23,6 @@ You can rebind keys with the in-game options (scroll to the bottom).
 
 There's a hard limit on how close to the edge you can build (I'll probably figure that out one of these days), so I wouldn't bother flattening the border mountains.
 
-Part of ECM.]],
+
+Part of Expanded Cheat Menu.]],
 })